Split roles by GUI&CLI, created crazymouse.py if statment to import corret module based on syntax
